Allow using Symfony 5.x, even with the 3.x branch of this bundle (#19)
This brings the pending changes from #17 and #18 to the 3.x branch of this bundle. Hopefully, it eases migration a bit.

Technically, it will be necessary to be at Symfony ^4.4 before the release containing these changes here can be obtained. But, being at Symfony 4.4 before attempting the migration to 5.0 should be the usual upgrade path.